Move handler to index.ts
