+++ pages/api/user/index.ts
@@ -1,6 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { ResponseUtil } from "@utils/responseUtil";
 import { getUsers } from "@database/user";
+import { getSession } from "next-auth/client";
+import { UserInput } from "@models/User";
+import { updateUser } from "@database/user";
 
 /**
  * handle controls the request made to the user resource
@@ -22,7 +25,25 @@ export default async function handle(
             break;
         }
         case "PUT": {
-            // TODO
+            const session = await getSession({ req });
+            const userId = session.id as string;
+            // TODO: add user role to session
+            const updatedUserData = {
+                id: userId,
+                firstName: req.body.firstName,
+                lastName: req.body.lastName,
+                role: req.body.role,
+                roleData: req.body.roleData,
+            } as UserInput;
+            const updatedUser = await updateUser(updatedUserData);
+            if (!updatedUser) {
+                ResponseUtil.returnBadRequest(
+                    res,
+                    `Error updating user with id ${userId}.`,
+                );
+                return;
+            }
+            ResponseUtil.returnOK(res, updatedUser);
             break;
         }
         case "DELETE": {
